+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - internal: precompiled gradle plugins
+- Update ErrorResponse to contain non-nullable `error` field
 
 ### Deprecated